final lab1

+++ AsymetricCryptoLab1/src/PseudoRandowNumberGenerators/LehmerGenerator/LehmerGenerator.java
@@ -4,7 +4,7 @@ import PseudoRandowNumberGenerators.Generator;
 
 public abstract class LehmerGenerator implements Generator {
     protected final long m=(long)1<<32;
-    protected final long a=(long)(((long)1<<16)+1);
+    protected final long a=(long)(((long)1<<16)+(long)1);
     protected final long c=119;
     protected long state;
